reservations filter, refactor

+++ app/src/BookReservations.App/ViewModels/RelationsViewModel.cs
@@ -1,11 +1,13 @@
 ﻿using BookReservations.Api.Client;
+using CommunityToolkit.Maui.Alerts;
+using CommunityToolkit.Maui.Core;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
 
 namespace BookReservations.App.ViewModels;
 
-[QueryProperty(nameof(RelationType), nameof(RelationType))]
+
 public partial class RelationsViewModel : ObservableObject, IViewModel
 {
     private readonly IApiClient apiClient;
@@ -30,13 +32,19 @@ public partial class RelationsViewModel : ObservableObject, IViewModel
         {
             return;
         }
-        await apiClient.DeleteRelationsAsync(ids);
+        var remove = await Shell.Current.DisplayAlert("Warning", "Are you sure?", "Yes", "No");
+        if (remove)
+        {
+            await apiClient.DeleteRelationsAsync(ids);
+            await Toast.Make("Relations was removed", ToastDuration.Long).Show();
+            await InitializeAsync();
+        }
     }
 
     [RelayCommand]
     private async Task GoToBookDetailAsync(int bookId)
     {
-        await Shell.Current.GoToAsync("//book/detail", new Dictionary<string, object>()
+        await Shell.Current.GoToAsync("//books/detail", new Dictionary<string, object>()
         {
             [nameof(BookViewModel.Id)] = bookId
         });
@@ -55,7 +63,7 @@ public partial class RelationsViewModel : ObservableObject, IViewModel
 
     public async Task InitializeAsync()
     {
-        RelationInfos.Clear();
+        RelationInfos = new();
         selectedRelationIds.Clear();
         var relations = (await apiClient.GetRelationsAsync(RelationType)).Result;
         foreach (var relation in relations)
